backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
sentence-transformers>=2.7.0
numpy>=1.26.0
openai>=1.3.0
firebase-admin>=6.2.0
beautifulsoup4>=4.12.0
requests>=2.31.0
PyPDF2>=3.0.0
python-docx>=1.1.0
lxml>=4.9.0

